Removing some unused includes

+++ Src/DasherCore/ButtonMode.cpp
@@ -8,7 +8,6 @@
 
 #include "DasherScreen.h"
 #include "DasherInterfaceBase.h"
-#include <valarray>
 
 
 using namespace Dasher;
@@ -113,7 +112,7 @@ void CButtonMode::SetupBoxes()
       double dMaxSize;
 
       if(dRatio == 1.0) {
-        dMaxSize = iDasherY / iForwardBoxes;
+        dMaxSize = static_cast<double>(iDasherY) / static_cast<double>(iForwardBoxes);
       }
       else {
         if(bEven)
